requirements.txt
altgraph==0.17.2
auto-py-to-exe==2.15.0
bottle==0.12.19
bottle-websocket==0.2.9
cffi==1.15.0
Eel==0.12.4
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
pefile==2021.9.3
Pillow==9.0.0
pycparser==2.21
pyinstaller==4.8
pyinstaller-hooks-contrib==2021.5
pyperclip==1.8.2
pyserial==3.5
pystray==0.19.2
pywin32-ctypes==0.2.0
six==1.16.0
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
